UpLife_api/requirements.txt
Django>=4.0,<5.0
Pillow>=8.0,<10.0
djangorestframework
psycopg2-binary
fastapi
uvicorn[standard]
django-cors-headers
dj-database-url

